Update init.lua and plugins
The commit updates the init.lua file of the project. It makes changes to the leader key, adds a useful plugin for keybinds, and removes commented code related to gitsigns.nvim.

Additionally, the commit adds a new file named lazy-lock.json. This file contains information about the specific versions and commits of various plugins used in the project.

The "custom/plugins/init.lua" file is also modified. The change adds a new plugin for a file explorer (nvim-tree.nvim) and configures it to toggle and focus the file explorer using specific key mappings.

Overall, these modifications improve the configuration and functionality of the Neovim setup.

Commit: Update init.lua and plugins